added a visual logout message for user when clicking logout button

+++ src/client/components/Logout.jsx
@@ -1,7 +1,11 @@
-function Logout() {
+function Logout({ setLogoutMessage }) {
 
     const handleSubmit = () => {
+      const token = localStorage.getItem('token')
+      if (token) {
         localStorage.clear()
+        setLogoutMessage('You have successfully logged out')
+      }
     }
 
     return (
